zshrc [proxy] Add file test on apt.conf

+++ zshrc
@@ -78,7 +78,11 @@ print_proxy () {
 	printf '\033[7mno_proxy\033[m\t%s\n' "$no_proxy"
 	printf '\033[7mNO_PROXY\033[m\t%s\n' "$NO_PROXY"
 	printf '\033[7m.gitconfig\033[m\t%s\n' "$(git config --global http.proxy)"
-	printf '\033[7mapt.conf\033[m\t%s\n' "$(grep --color=never '^Acquire::http::Proxy' /etc/apt/apt.conf)"
+	if test -f /etc/apt/apt.conf
+	then
+		printf '\033[7mapt.conf\033[m\t%s\n' \
+			"$(grep --color=never '^Acquire::http::Proxy' /etc/apt/apt.conf)"
+	fi
 }
 
 proxy () {
@@ -96,9 +100,12 @@ proxy () {
 			export no_proxy=bull.fr,ao-srv.com
 			export NO_PROXY=$no_proxy
 			git config --global http.proxy "$http_proxy"
-			grep '^Acquire::http::Proxy' /etc/apt/apt.conf > /dev/null \
-				|| echo "Acquire::http::Proxy \"$http_proxy\";" \
-					| sudo tee -a /etc/apt/apt.conf > /dev/null
+			if test -f /etc/apt/apt.conf
+			then
+				grep '^Acquire::http::Proxy' /etc/apt/apt.conf > /dev/null \
+					|| echo "Acquire::http::Proxy \"$http_proxy\";" \
+						| sudo tee -a /etc/apt/apt.conf > /dev/null
+			fi
 			print_proxy
 			;;
 		off )
@@ -109,8 +116,11 @@ proxy () {
 			export no_proxy=
 			export NO_PROXY=$no_proxy
 			git config --global --unset http.proxy
-			grep '^Acquire::http::Proxy' /etc/apt/apt.conf > /dev/null \
-				&& sudo sed -i '/^Acquire::http::Proxy/d' /etc/apt/apt.conf
+			if test -f /etc/apt/apt.conf
+			then
+				grep '^Acquire::http::Proxy' /etc/apt/apt.conf > /dev/null \
+					&& sudo sed -i '/^Acquire::http::Proxy/d' /etc/apt/apt.conf
+			fi
 			print_proxy
 			;;
 		* )
